4-Context-Aware-RAG-Assistant/requirements.txt
fastapi
uvicorn
openai
chromadb
tiktoken
python-dotenv
